queryDSL과 JPA를 병행하여 Repository 구현.

+++ backend/src/main/kotlin/com/back/backend/domain/transactions/repository/AccountTransactionRepository.kt
@@ -3,7 +3,7 @@ package com.back.backend.domain.transactions.repository
 import com.back.backend.domain.transactions.entity.AccountTransaction
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface AccountTransactionRepository : JpaRepository<AccountTransaction, Long> {
+interface AccountTransactionRepository : JpaRepository<AccountTransaction, Int> {
     fun findByAccountId(accountId: Int): List<AccountTransaction>
     fun findByAccountIdIn(accountIds: List<Int>): List<AccountTransaction>
 }
